refactor : 연산 중에 오버플로우 예외 처리
